Set up the Parameter Dynamics

The objects presented in this section are usefull to set up the dynamics of the parameter system and can be used to specify how there dynamics should be updated according to the past trajectory.

FiveBoxModel.ParameterDynamics — Type

ParameterDynamics

A mutable struct which encodes the RHS of the dynamics of the parameter space

Fields

f!::Function: Describes the deterministic dynamics in parameter space with f! of the from f!(du,u,t_0,t) with length(du) = length(u) = 26 and u contains all the salinities together with all the parameter values. Since the system in run in sub-time-intervals the current time is given by t+t_0 which which is relevent it the system is non autonomous. Also note that all times are given in the unit of years in order to avoid very large numbers for t.

g!::Function: Describes the stochastic dynamics in parameter space and is of the same form as f!

noise_rate_prototype::Matrix{Float64}: Is some Matrix with the same dimensions as the du argument in g! which is later needed to set up the SEDProblem from the DifferentialEquations package. If noise_rate_prototype === nothing then diagonal noise is used and du has the form of a Vector. Note that noise_rate_prototype must be the same must have the same sizes for corresponding instances of the ParameterDynamics and PhaseDynamics strucs.

If only f! and g! are given then update will be initialized to update = nothing

FiveBoxModel.ParameterDynamics — Method

ParameterDynamics(f!::Function, 
                    g!::Function;   
                    noise_rate_prototype::Union{Matrix{Float64},Nothing} = nothing)

Alternative instantiation which returns an object of ParameterDynamics.

Presets

If either f! or g! as arguments of ParameterDynamics are supposed to be zero one can respectively use the presets for these functions.

FiveBoxModel.no_deterministic_param_dynamics! — Function

function no_deterministic_param_dynamics!(du, u, t_0, t)

This function encodes the deterministic part of the ODEs describing the dynamics of the parameters in the case where the parameters do not change deterministically with time. This can be used as argument for f! when ceating an instance of the type ParameterDynamics.

FiveBoxModel.no_stochastic_param_dynamics! — Function

no_stochastic_param_dynamics!(du, u, t_0, t)

This function encodes the stochastic part of the ODEs describing the dynamics of the parameters in the case where this part is zero. This can be used as argument for g! when ceating an instance of the type ParameterDynamics.

If both are zero simultaneously there also is a preset instance of an ParameterDynamics object

FiveBoxModel.fixedParameters — Constant

fixedParameters

This is a preset of the type ParameterDynamics where the parameters do not change during the evolution of the system.

Examples

Hosing experiment

For this example we only want deterministic parameter changes which occur as follows and there is no stochastic part to the parameter dynamics. We call the deterministic part hosing_experiment!. This can then be used to create a respective instand of the type ParameterDynamics. Since we do not want f! and g! to be updated during the integration process we instanciate the object in the following way

julia> function hosing_experiment!(du, u, t_0, t)
           H_1 = 5e-4
           du[get_index["parameters"]] .= 0.
           if t+t_0 <= 2000
               sign = 1
           else
               sign = -1
           end
           du[get_index["F_N"]] = sign*H_1*0.194
           du[get_index["F_S"]] = -sign*H_1*0.226
           du[get_index["F_T"]] = sign*H_1*0.597
           du[get_index["F_IP"]] = -sign*H_1*0.565
       end
hosing_experiment! (generic function with 1 method)

julia> using FiveBoxModel

julia> hosingExperiment = ParameterDynamics(hosing_experiment!,no_stochastic_param_dynamics!)
ParameterDynamics(hosing_experiment!, FiveBoxModel.no_stochastic_param_dynamics!, nothing, nothing)