Set up the Phase Dynamics

The objects presented in this section are usefull to set up the dynamics of the phase system and can be used to specify how these dynamics should be updated according to the past trajectory.

FiveBoxModel.PhaseDynamics — Type

PhaseDynamics

A mutable struct which encodes the RHS of the dynamics of the phase space

Fields

f!::Function: Describes the deterministic dynamics in phase space with f! of the from f!(du,u,t_0,t) with length(du) = length(u) = 26 and u contains all the salinities together with all the parameter values. Since the system in run in sub-time-intervals the current time is given by t+t_0 which which is relevent it the system is non autonomous. Also note that all times are given in the unit of years in order to avoid very large numbers for t.

g!::Function: Describes the stochastic dynamics in phase space and is of the same form as f!.

noise_rate_prototype::Matrix{Float64}: Is some Matrix with the same dimensions as the du argument in g! which is later needed to set up the SEDProblem from the DifferentialEquations package. If noise_rate_prototype === nothing then diagonal noise is used and du has the form of a Vector. Note that noise_rate_prototype must be the same must have the same sizes for corresponding instances of the ParameterDynamics and PhaseDynamics strucs.

If only f! and g! are given then update will be initialized to update = nothing

FiveBoxModel.PhaseDynamics — Method

PhaseDynamics(f!::Function, 
                g!::Function;   
                noise_rate_prototype::Union{Matrix{Float64},Nothing} = nothing)

Alternative instantiation which returns an object of ParameterDynamics.

Presets

If f! is supposed to be given by the RHS of the deterministic salinity evolution as described in the paper or g! as argument of PhaseDynamics is supposed to be zero, one can respectively use the presets for these functions.

FiveBoxModel.deterministic_salinity_dynamics! — Function

deterministic_salinity_dynamics!(du, u, t_0, t)

This function encodes the deterministic part of the ODEs describing the dynamics of the parameters in the case where the deterministic evolution of the salinities in time is given by the respective equations of motion (the modified versions). This can be used as argument for f! when ceating an instance of the type ParameterDynamics.

FiveBoxModel.original_deterministic_salinity_dynamics! — Function

deterministic_salinity_dynamics!(du, u, t_0, t)

This function encodes the deterministic part of the ODEs describing the dynamics of the parameters in the case where the deterministic evolution of the salinities in time is given by the respective equations of motion (the original equations as given in the paper). This can be used as argument for f! when ceating an instance of the type ParameterDynamics.

FiveBoxModel.no_stochstic_salinity_dynamics! — Function

no_stochastic_salinity_dynamics!(du, u, t_0, t)

This function encodes the stochastic part of the ODEs describing the dynamics of the salinities in the case where this part is zero. This can be used as argument for g! when ceating an instance of the type ParameterDynamics.

For certain combinations of the above cases in a situation where no updates of f! and g! are intended. The following presets might be usefull

FiveBoxModel.DeterministicPhaseDynamics — Constant

DeterministicPhaseDynamics

This is a preset of the type PhaseDynamics where the salinity evolutions is fully deterministic and given by the modified equations.

FiveBoxModel.OriginalDeterministicPhaseDynamics — Constant

OriginalDeterministicPhaseDynamics

This is a preset of the type PhaseDynamics where the salinity evolutions is fully deterministic and given by the original equations from the paper.

RHS of the ODEs describing the salinity evolution

For a more detailed access to the RHSs of the ODEs describing the salinity evolution the following methods might be usefull

FiveBoxModel.RHS_module.rhs_S_N — Function

rhs_S_N(arr::Array{Float64}; original::Bool = false)

Arguments

arr::Array{Float64}: Array of length 26 containing the current state of the system at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_N

rhs_S_N(sys::System; original::Bool = false)

Arguments

sys::System: system containing the current state at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_N

FiveBoxModel.RHS_module.rhs_S_T — Function

rhs_S_T(arr::Array{Float64}; original::Bool = false)

Arguments

arr::Array{Float64}: Array of length 26 containing the current state of the system at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_T

rhs_S_T(sys::System; original::Bool = false)

Arguments

sys::System: system containing the current state at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_T

FiveBoxModel.RHS_module.rhs_S_S — Function

rhs_S_S(arr::Array{Float64}; original::Bool = false)

Arguments

arr::Array{Float64}: Array of length 26 containing the current state of the system at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_S

rhs_S_S(sys::System; original::Bool = false)

Arguments

sys::System: system containing the current state at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_S

FiveBoxModel.RHS_module.rhs_S_IP — Function

rhs_S_IP(arr::Array{Float64}; original::Bool = false)

Arguments

arr::Array{Float64}: Array of length 26 containing the current state of the system at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_IP

rhs_S_IP(sys::System; original::Bool = false)

Arguments

sys::System: system containing the current state at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_IP

FiveBoxModel.RHS_module.rhs_S_B — Function

rhs_S_B(arr::Array{Float64}; original::Bool = false)

Arguments

arr::Array{Float64}: Array of length 26 containing the current state of the system at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_B

rhs_S_B(sys::System; original::Bool = false)

Arguments

sys::System: system containing the current state at which the RHS of the ODE is to be determined

original::Bool: If false then the modified version of the RHS of the ODE is returned and if true then the original version as presented in the paper is used

Returns

The RHS of the ODE describing the rate of change of the varialbe S_B

FiveBoxModel.RHS_module.get_q — Function

get_q(arr::Array{Float64})

Returns the value of q for an array describing a given system. To this end the euqations q = λ[α(Tₛ-Tₙ)+β(Sₙ-Sₛ)] and Tₙ = μ⋅q+T₀ where solved for q

get_q(sys::System)

Returns the value of q of a given system. To this end the euqations q = λ[α(Tₛ-Tₙ)+β(Sₙ-Sₛ)] and Tₙ = μ⋅q+T₀ where solved for q