System

An object of the type System contains the values of all five salinities in form of an Array and all parameter values as an Object of type ParameterSet, i.e. a system contains all the information of a given system at a single point in time.

FiveBoxModel.RHS_module.System — Type

System

Mutalbe struct which carries the current sylinity values as well as the parameter values of a given system

Fields

salinities::Array: Array of size (5,) of the form [SN,ST,SS,SIP,S_B] where each salinity is given in units of psu (Any five element Array is reshaped into this form at initialization of the stuct)

params::ParameterSet: Contains the current parameters describing the system

An alternative instantiation is given by

FiveBoxModel.RHS_module.System — Method

System(;S_N::Number,
        S_T::Number,
        S_S::Number,
        S_IP::Number,
        S_B::Number,
        params::ParameterSet)

Returns an element of the type System, of the form System([S_N,S_T,S_S,S_IP,S_B],params) (All salinities must be given in units of psu)

An instance of ParameterSet which is needed to create a System is of the following form

FiveBoxModel.RHS_module.ParameterSet — Type

ParameterSet

Stuct whos fields describe all Parameters of a given System. The fields are to be given as keyword arguments.

Fields

λ::Number: [λ] = m^6 * kg^{-1} * s^{-1} * 10^7

α::Number: [α] = kg * m^{-3} * K^{-1}

β::Number: [β] = kg * m^{-3} * psu^{-1}

γ::Number: [γ] = 1

η::Number: [η] = m^3 * s^{-1} * 10^6

μ::Number: [μ] = °C * m^{-3} * s * 10^{-8}

V_N::Number: [V_N] = m^3 * 10^16

V_S::Number: [V_S] = m^3 * 10^16

V_T::Number: [V_T] = m^3 * 10^16

V_IP::Number: [V_IP] = m^3 * 10^16

V_B::Number: [V_B] = m^3 * 10^16

K_N::Number: [K_N] = m^3 * s^{-1} * 10^6

K_S::Number: [K_S] = m^3 * s^{-1} * 10^6

K_IP::Number: [K_IP] = m^3 * s^{-1} * 10^6

F_N::Number: [F_N] = m^3 * s^{-1} * 10^6

F_S::Number: [F_S] = m^3 * s^{-1} * 10^6

F_T::Number: [F_T] = m^3 * s^{-1} * 10^6

F_IP::Number: [F_IP] = m^3 * s^{-1} * 10^6

S_0::Number: [S_0] = psu

T_0::Number: [T_0] = °C

T_S::Number: [T_S] = °C

Both instances of System and ParameterSet can be updated respectively via

FiveBoxModel.RHS_module.updateSystem! — Function

updateSystem!(sys::System, arr::Array{Float64})

Updates the values in the System according to the values given in the array.

FiveBoxModel.RHS_module.updateParameterSet! — Function

updateParameterSet!(params::ParameterSet, arr::Array)

Updates the given ParameterSet with the values given in the Array.

Presets

For the FAMOUS simulation there are presets for both the initial parameter values and the salinities

FiveBoxModel.RHS_module.FAMOUSparams — Constant

FAMOUSparams

A ParameterSet containing the parameters as measured from the FAMOUS simulation.

FiveBoxModel.RHS_module.FAMOUS — Constant

FAMOUS

Preset of a instance of type System which carries the initial values from the FAMOUS experiment

To Array conversion

Often it is more convenient to work with arrays rather then with systems. The following method allows a conversion from a System into an Array

FiveBoxModel.RHS_module.toArray — Function

ToArray(sys::System)

Converts the data contained in the System into an array

To obtain certain data concerning a given variable of interest from an Array, one can use the function get_VariableName(arr::Array) where "VariableName" must be replaced with the name of the respective variable of interest. In case arr represents a full trajectory, get_VariableName will slice out the respective part of the trajectory describing the evolution of the variable of interest. The same function also works when applied to systems, i.e. get_VariableName(sys::System). An alternative method would be to extract the index of the variable of interest via the following dictionary

FiveBoxModel.RHS_module.get_index — Constant

get_index

When converting objects of type System into Arrays it might be confusiong to ceep track to which entry in the array encodes which variable. The dictionary get_index encodes exactly this information where the keys are the names of the respective variables.

This is also illustrated in the following example

julia> using FiveBoxModel

julia> get_S_N(FAMOUS)
35.088

julia> arr = toArray(FAMOUS)
26-element Vector{Float64}:
  35.088
   ⋮
   5.571

julia> get_S_N(arr)
35.088

julia> arr[get_index["S_N"]]
35.088